+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to the Path Server will be documented in this file.
 ## [Unreleased]
 ## Fixed
 - **Core**: Fixed an issue where path in markdown quote and html block may not extract correctly.
+- **Core**: Fixed an issue where raw string in rust may not extract correctly.
 
 ## [0.5.1] - 2026-03-12
 ### Added